Move Mask foreign impls and tests into submodule structure
